requirements.txt
requests>=2.28.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0
fastmcp>=2.5.1
python-dotenv>=1.0.0
uvicorn>=0.22.0
httpx
